+++ debug/example/recyclerview/recycle_gridview.js
@@ -16,11 +16,13 @@ const onReady = () => {
     const gridView = new data[1].GridView();
     // 출력
     document.querySelector('#left').innerHTML = layout.render({});
-    return gridView;
+    // response
+    const reponse = { data: Array.isArray(gridData) ? gridData : [gridData], view: gridView };
+    return reponse;
   }).
-  then((gridView) => {
+  then((reponse) => {
     // RecyclerView 인스턴스 생성 및 초기화
-    const adapter = new SimpleAdapter(gridData, gridView);
+    const adapter = new SimpleAdapter(reponse.data, reponse.view);
     const recyclerView = new RecyclerView('#gridview', adapter, {
       itemCount: 8,
       clientRectHeight: 256,