+++ debug/example/recyclerview/recycle_tableview.js
@@ -5,9 +5,9 @@ const onReady = () => {
   let page = 1;
   function doList(page) {
     // 데이터
-    const data = [];
+    const tableData = [];
     for (let i = 0; i <= 6000; i++) {
-      data.push({ id: i, gid: (i + 1) / 10, title: `Item${page} ${i}` });
+      tableData.push({ id: i, gid: (i + 1) / 10, title: `Item${page} ${i}` });
     }
     Promise.all([
     new AsyncTask().doImport(new URL(`../recyclerview/tpl/layout_table.js`, import.meta.url).href),
@@ -18,11 +18,13 @@ const onReady = () => {
       const tableView = new data[1].TableView();
       // 출력
       document.querySelector('#left').innerHTML = layout.render({});
-      return tableView;
+      // response
+      const reponse = { data: Array.isArray(tableData) ? tableData : [tableData], view: tableView };
+      return reponse;
     }).
-    then((tableView) => {
+    then((reponse) => {
       // RecyclerView 인스턴스 생성 및 초기화
-      const adapter = new SimpleAdapter(data, tableView);
+      const adapter = new SimpleAdapter(reponse.data, reponse.view);
       const recyclerView = new RecyclerView('#tableview', adapter, { itemCount: 10, clientRectHeight: 50, scrollCapture: '#fvue--layout--main' });
       recyclerView.addEventListener('click', '.item', (el) => {
         const id = el.dataset.id;